Todo-TypeScript
Providing Items Count

provided counts of the total number of TodoItem objects, the number that are complete and the number still outstanding.
added a method to the TodoCollection class that returns  an object that describes the items in the collection.
updated the index.ts file so that the number of incomplete items is displayed to the user.